% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_user_code_call.R
\name{last_user_code_call}
\alias{last_user_code_call}
\title{Get last call in user code, excluding comments}
\usage{
last_user_code_call(.user_code)
}
\arguments{
\item{.user_code}{User-provided code in a learnr tutorial}
}
\value{
The last string provided by the user in their code
}
\description{
Get last call in user code, excluding comments
}
