% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tutorials}
\alias{list_tutorials}
\title{List all tutorials}
\usage{
list_tutorials()
}
\value{
Character vector
}
\description{
Lists all tutorials (Rmd files) in joholearnr
}
\examples{
list_tutorials()
}
