% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snapshot_tutorials}
\alias{snapshot_tutorials}
\title{Snapshot all tutorial files}
\usage{
snapshot_tutorials(dir = "./tutorials")
}
\arguments{
\item{dir}{Directory containing the tutorials. Each should be in its own
subdirectory; see \code{system.file("tutorials", package = "joholearnr")}}
}
\value{
Nothing. Externally, the renv files will be written to disk.
}
\description{
Create an renv.lock file for each tutorial
}
\examples{
if (interactive()) {
  tutorials_folder <- system.file("tutorials", package = "joholearnr")
  temp_tutorials_folder <- fs::path(tempdir(), "tutorials")
  fs::dir_copy(tutorials_folder, temp_tutorials_folder, overwrite = TRUE)
  snapshot_tutorials(temp_tutorials_folder)
  fs::dir_delete(temp_tutorials_folder)
}
}
