% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{this_is_a_really_long_name}
\alias{this_is_a_really_long_name}
\title{Object with a long name}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
this_is_a_really_long_name
}
\description{
This is provided only for practicing using the tab key to autocomplete a long
name
}
\keyword{datasets}
