% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_all_tutorials}
\alias{write_all_tutorials}
\title{Write out all tutorial files}
\usage{
write_all_tutorials(dir = ".", delete_html = TRUE)
}
\arguments{
\item{dir}{Directory to write the files.}

\item{delete_html}{Logical; should HTML files in the tutorials directory
be deleted after copying?}
}
\value{
Nothing. Externally, the file will be written to disk. (Subfolders
may be used)
}
\description{
Write out all tutorial files (Rmd) file to disk
}
\examples{
target_dir <- fs::path(tempdir(), "tutorials")
write_all_tutorials(target_dir)
fs::dir_ls(target_dir)
fs::dir_delete(target_dir)
}
