% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_tutorial}
\alias{write_tutorial}
\title{Write out a tutorial file}
\usage{
write_tutorial(tutorial, dir = ".")
}
\arguments{
\item{tutorial}{Name of the tutorial file. The .Rmd extension may be omitted.}

\item{dir}{Directory to write the file.}
}
\value{
Nothing. Externally, the file will be written to disk.
}
\description{
Write out a tutorial (Rmd) file to disk
}
\examples{
write_tutorial("00-learnr", tempdir())
list.files(tempdir())
}
